08/requirements.txt
Flask==3.1.2
flask-marshmallow==1.3.0
Flask-Migrate
Flask-SQLAlchemy
flask-socketio
marshmallow-sqlalchemy